Split flies for functions and classes
